refactor: 清理冗余代码并全面校正文档与前端类型问题
- 后端代码冗余清理（仅确定冗余）
  - auth handler 中移除冗余闭包调用
  - document lock 中移除无意义同型类型转换
  - database tests 中改为结构体初始化，避免 Default 后字段重赋值

- 文档全量校正（以当前源码为准）
  - 重写入口文档：README、QUICKSTART、docs/README
  - 更新模块文档：core/crypto/api/web/auth/storage/database
  - 更新流程与环境文档：docs/flows、docs/environment
  - 历史文档保留正文并追加“当前状态更正”：CHANGELOG、TODO、TEST_REPORT
  - 更新 CLAUDE 与 web/README 快照说明

- 前端 lint 质量债务清理
  - 消除 no-explicit-any / no-unused-vars / hooks 依赖问题
  - 修复 PreviewModal 的 setState-in-effect 问题
  - 收紧 API 与 Store 类型定义，完善错误处理类型分支
  - 修复 DocumentsPage 可选字段导致的构建类型错误

- 验证
  - cargo check --workspace 通过
  - cargo clippy --workspace --all-targets -- -D warnings 通过
  - web: npm run lint 与 npm run build 均通过

+++ crates/rustcloud-api/src/services/document_lock.rs
@@ -118,7 +118,7 @@ impl DocumentLockManager {
         }
 
         // 延长过期时间
-        let _: bool = conn.expire(&key, LOCK_TTL_SECONDS as i64).await?;
+        let _: bool = conn.expire(&key, LOCK_TTL_SECONDS).await?;
 
         tracing::debug!(
             doc_id = %doc_id,
